[Silver III] Title: 피보나치 수의 확장, Time: 116 ms, Memory: 18048 KB -BaekjoonHub

+++ 백준/Silver/1788. 피보나치 수의 확장/피보나치 수의 확장.java
@@ -10,7 +10,7 @@ class Main{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int N = Integer.parseInt(br.readLine());
 		int len = Math.abs(N);
-		long dp[] = new long[len+2];	// 양수
+		int dp[] = new int[len+2];	// 양수
 		dp[1] = 1;
 		
 		for(int i=2; i<=len; i++)
